requirements.txt
autopep8==2.0.0
beautifulsoup4==4.11.1
cassandra-driver==3.25.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
geomet==0.2.1.post1
idna==3.4
install==1.3.5
mariadb==1.0.11
mysql-connector-python==8.0.31
protobuf==3.20.1
pycodestyle==2.9.1
pycoingecko==3.1.0
python-coinmarketcap==0.5
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.12
